Directly change the semantic exit to hostio require(0) in zkWASM
Arguement is:
In a normal exit, goroutine set var `PAUSE` to nozero to end the current
main thread and an accident exit, the runtime `exit` is chosen.

So our change for a normal hostio exit, we set PAUSE to 1 to directly exit.
For WASM, the exit now is only for abnormal exit, which includes panic.
So we use zkWASM hostio `require` with input 0 marks it is abnormal and
needs exit